requirements.txt
annotated-types==0.6.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
python-vlc-http==0.0.11
requests==2.31.0
typing_extensions==4.11.0
urllib3==2.2.1
xmltodict==0.13.0
